refactor: replace double with decimal for monetary values and improve code readability
Changed double to decimal for monetary properties to ensure precision in financial calculations. Improved code readability by refactoring LINQ queries, removing redundant code, and aligning with best practices. Updated method names in IConfigurationService for consistency.

+++ Backend/Application/Dashboard/Dtos/DashboardCard.cs
@@ -2,11 +2,11 @@ namespace Application.Dashboard.Dtos;
 
 public class DashboardCard
 {
-    public double Revenue { get; set; }
+    public decimal Revenue { get; set; }
     public double RevenueDiff { get; set; }
     public int Order { get; set; }
     public double OrderDiff { get; set; }
-    public double AverageOrderValue { get; set; }
+    public decimal AverageOrderValue { get; set; }
     public double AverageOrderValueDiff { get; set; }
     public int NewUser { get; set; }
     public double NewUserDiff { get; set; }
@@ -14,7 +14,7 @@ public class DashboardCard
 
 public class RevenueCard
 {
-    public double? Revenue { get; set; }
+    public decimal? Revenue { get; set; }
     public double? RevenueDiff { get; set; }
 }
 
@@ -26,7 +26,7 @@ public class OrderCard
 
 public class AverageOrderValueCard
 {
-    public double? AverageOrderValue { get; set; }
+    public decimal? AverageOrderValue { get; set; }
     public double? AverageOrderValueDiff { get; set; }
 }
